Implement 'actions' into Tasks
The Task model gets a new field called 'actions' which contains key, value pairs in JSON format. Keys will be displayed as button labels and values must be valid url names. If 'actions' field is not empty, the home page for a given task will loop through the 'actions' JSON and display each action as a button.