Fix bug in deployer.py
When sequencing the commands, the script was checking if migrations folder existed for an app. Because no commands are actually run at this point, the script was assuming that these folders would be there so it did not queue the migration creation commands. However, if user requests to reset DB, migration folders are removed, but migrations are not created because the creation commands were never queued.

This is fixed by conditioning creation of migration folders not only on whether they can be found in the system, but also on whether these folders were queued for deletion by the deployer.